Improve the Pager subprocess without writing to a file first
The 'trick' is that it should communicate with the TTY. That's described in https://jonisalonen.com/2012/runtime-exec-with-unix-console-programs/

    Process p = Runtime.getRuntime().exec(new String[]{"sh", "-c", "less -FRX >/dev/tty"});
    var pw = new PrintWriter(p.getOutputStream());
    for (int i = 0; i < 100; i++) {
        if (pw.checkError()) {
            break;
        }
        pw.println("hello " + i);
    }
    pw.close();
    System.out.println("=> " + p.waitFor());

Also: https://stackoverflow.com/questions/29733038/running-interactive-shell-program-in-java

It also seems that the `write` to the process `PrintStream` is
blocking/buffered when writing a lot of data to the output stream into the
`less` program, until it is read by less. That is good, for example when
implementing `git log` so it won't read the entire database if it's not even
read.

By using `stdout.checkError()`  it checks if the CommandContext.stdout is still
open and that it makes sense to write to it. This is really helpful if the
program is piped into another program, which doesn't need more data.

And example of such a program is head:

    $ gikt paged --times 10000000 | head
    $ gikt log | grep "commit msg"