16.2.4 Filtering by changed paths
Add TreeEntry.key for easy creating a sub-filter of the PathFilter with PathFilter#join

With these changes, paths from the workspace can be combined into a PathFilter.
The filter is used to filter the TreeDiff between two commits.
If the diff is empty, the RevList item is marked with TREESAME, and is filtered out from the logs